refactor(ui): extract formatTranscriptDate to shared lib/format
Consolidates duplicate date formatting from TranscriptRow and
TranscriptViewer into a single shared utility.
