+++ cloud/apps/web/src/lib/format.ts
@@ -3,6 +3,26 @@
  */
 
 /**
+ * Format a transcript's createdAt timestamp for display in lists and detail views.
+ * Always shows MM/DD/YYYY HH:MM AM/PM Pacific.
+ */
+export function formatTranscriptDate(dateString: string): string {
+  const date = new Date(dateString);
+  const datePart = date.toLocaleDateString('en-US', {
+    month: '2-digit',
+    day: '2-digit',
+    year: 'numeric',
+    timeZone: 'America/Los_Angeles',
+  });
+  const timePart = date.toLocaleTimeString('en-US', {
+    hour: '2-digit',
+    minute: '2-digit',
+    timeZone: 'America/Los_Angeles',
+  });
+  return `${datePart} ${timePart} Pacific`;
+}
+
+/**
  * Format a run's display name.
  * If the run has a custom name, use it.
  * Otherwise, generate an algorithmic name from the definition name and date.
